+++ PromotionEngineBL/ICartBL.cs
@@ -5,6 +5,6 @@ namespace PromotionEngineBL
 {
     public interface ICartBL
     {
-        public int GetCartPrice(List<CartProduct> cartProducts);
+        public decimal GetCartPrice(List<CartProduct> cartProducts);
     }
 }
